backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
Pillow
PyMuPDF
python-dotenv
google-generativeai
